requirements.txt
# Core API Framework
fastapi==0.135.3
uvicorn[standard]==0.44.0
starlette==1.0.0

# AI/LLM Integration
groq==0.37.1
langchain==1.2.15
langchain-core==1.2.30
langchain-groq==1.1.2

# Data Validation
pydantic==2.13.1
pydantic-core==2.46.1
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1

# Environment & Config
python-dotenv==1.2.2

# HTTP & Network
httpx==0.28.1
httpcore==1.0.9
requests==2.33.1
urllib3==2.6.3

# Utilities & Dependencies
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
idna==3.11
sniffio==1.3.1
typing_extensions==4.15.0

# Performance & Serialization (opcional pero recomendado)
uvloop==0.22.1
orjson==3.11.8
httptools==0.7.1

# Required by dependencies
annotated-types==0.7.0
tenacity==9.1.4
